refactor: Move framework form wizard to shared location
This router will be used for future assessment so is being moved
to a place in the codebase from where it can be shared more easily.
